# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Appliance Service Center Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List service jobs
      parameters:
        - name: customerId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
        - name: paymentStatus
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceJob"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a service job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceJobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceJob"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceJob"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceJobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceJob"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /jobs/{id}/payment:
    patch:
      operationId: updateJobPayment
      tags: [jobs]
      summary: Update payment status of a job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceJob"
  /jobs/recent:
    get:
      operationId: listRecentJobs
      tags: [jobs]
      summary: Get recent jobs (dashboard feed)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Recent jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceJob"

  # ── Highlights ──────────────────────────────────────────────
...
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List reminders
      parameters:
        - name: customerId
          in: query
          schema: { type: integer }
        - name: isActive
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Users ──────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Settings ──────────────────────────────────────────────
...
components:
  schemas:
...
    ServiceJob:
      type: object
      required: [id, jobNumber, customerId, status, paymentStatus, createdAt]
      properties:
        id: { type: integer }
        jobNumber: { type: string }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        customerPhone: { type: ["string", "null"] }
        applianceId: { type: ["integer", "null"] }
        applianceType: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        paymentStatus:
          type: string
          enum: [unpaid, partial, paid]
        amount: { type: number }
        paidAmount: { type: number }
        technicianName: { type: ["string", "null"] }
        scheduledDate: { type: ["string", "null"] }
        completedDate: { type: ["string", "null"] }
        isHighlighted: { type: boolean }
        isNumberHighlighted: { type: boolean }
        highlights:
          type: array
          items:
            $ref: "#/components/schemas/Highlight"
        createdAt: { type: string }
        updatedAt: { type: string }
    ServiceJobInput:
      type: object
      required: [customerId]
      properties:
        customerId: { type: integer }
        applianceId: { type: integer }
        description: { type: string }
        status: { type: string }
        paymentStatus: { type: string }
        amount: { type: number }
        paidAmount: { type: number }
        technicianName: { type: string }
        scheduledDate: { type: string }
        isHighlighted: { type: boolean }
        isNumberHighlighted: { type: boolean }
    ServiceJobUpdate:
      type: object
      properties:
        applianceId: { type: integer }
        description: { type: string }
        status: { type: string }
        paymentStatus: { type: string }
        amount: { type: number }
        paidAmount: { type: number }
        technicianName: { type: string }
        scheduledDate: { type: string }
        completedDate: { type: string }
        isHighlighted: { type: boolean }
        isNumberHighlighted: { type: boolean }
    PaymentUpdate:
      type: object
      required: [paymentStatus]
      properties:
        paymentStatus:
          type: string
          enum: [unpaid, partial, paid]
        paidAmount: { type: number }

    # ── Highlight ──────────────────────────────────────────────
    Highlight:
      type: object
      required: [id, label, color, createdAt]
      properties:
        id: { type: integer }
        jobId: { type: ["integer", "null"] }
        customerId: { type: ["integer", "null"] }
        label: { type: string }
        color: { type: string }
        captionSize: { type: number }
        isNumbered: { type: boolean }
        isTicked: { type: boolean }
        zoomLevel: { type: number }
        createdAt: { type: string }
...
    Reminder:
      type: object
      required: [id, title, reminderAt, isActive, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        jobId: { type: ["integer", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        reminderAt: { type: string }
        isActive: { type: boolean }
        isTriggered: { type: boolean }
        createdAt: { type: string }
    ReminderInput:
      type: object
      required: [title, reminderAt]
      properties:
        customerId: { type: integer }
        jobId: { type: integer }
        title: { type: string }
        description: { type: string }
        reminderAt: { type: string }
        isActive: { type: boolean }
    ReminderUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        reminderAt: { type: string }
        isActive: { type: boolean }
        isTriggered: { type: boolean }

    # ── AppUser ──────────────────────────────────────────────
    AppUser:
      type: object
      required: [id, name, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        role:
          type: string
          enum: [admin, technician, viewer]
        permissions:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
    AppUserInput:
      type: object
      required: [name, role]
      properties:
        name: { type: string }
        email: { type: string }
        role: { type: string }
        permissions:
          type: array
          items: { type: string }
    AppUserUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        role: { type: string }
        permissions:
          type: array
          items: { type: string }
        isActive: { type: boolean }

    # ── AppSettings ──────────────────────────────────────────────
...