# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Appliance Service Center Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Reminder {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  jobId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  reminderAt: string;
  isActive: boolean;
  isTriggered?: boolean;
  createdAt: string;
}